requirements.txt
google-genai>=0.2.0
faiss-cpu
numpy
tqdm
python-dotenv
tenacity
sentence-transformers
rank-bm25
